'popravke2'

+++ plugins/Route/Save.php
@@ -35,7 +35,7 @@ $out = array(
 if ($_REQUEST['TopRoute']==1) $result = $dbT->RunQuery("INSERT IGNORE INTO `v4_TopRoutes`(`TopRouteID`) VALUES (".$topRouteID.")");
 else $result = $dbT->RunQuery("DELETE FROM `v4_TopRoutes` WHERE `TopRouteID`=".$topRouteID);
 
-$result = $dbT->RunQuery("DELETE FROM `v4_RoutesTerminals` WHERE `RouteID`=".$keyValue);
+if ($keyName != '' and $keyValue != '') $result = $dbT->RunQuery("DELETE FROM `v4_RoutesTerminals` WHERE `RouteID`=".$keyValue);
 // varijanta iz select box-a
 $terminalID=$_REQUEST['TerminalID'];
 /*$result2 = $dbT->RunQuery("SELECT count(*) as count from v4_Terminals WHERE TerminalID=".$_REQUEST['FromID']);
@@ -87,7 +87,7 @@ else {
 	}
 }*/
 if ($terminalID>0) {
-	$dbT->RunQuery("INSERT INTO `v4_RoutesTerminals`(`RouteID`, `TerminalID`) VALUES (".$keyValue.",".$terminalID.")");					
+	$dbT->RunQuery("INSERT INTO `v4_RoutesTerminals`(`RouteID`, `TerminalID`) VALUES (".$topRouteID.",".$terminalID.")");					
 }	
 # send output back
 $output = json_encode($out);
